embedded: remove old mixer
